dev refactored
