+++ .auth/user.json
@@ -2,8 +2,8 @@
   "cookies": [
     {
       "name": "authjs.csrf-token",
-      "value": "ba06d84c706f616e598968cc376e1a4304c0a71db1305ddfd952924156d49d1a%7C4711740cd25e3f1d882f8126e7e317df634b09b5829e1f0f57808b8546fe0413",
-      "domain": "localhost",
+      "value": "a867f201279b76f9927eb5d732b4e118e2a36d20b770bc5dd3f02a25c0fb73e7%7C9c2402f31e5574b29c5d06458c10e219f02a1c327959b009ea31aadb69e63adb",
+      "domain": "127.0.0.1",
       "path": "/",
       "expires": -1,
       "httpOnly": true,
@@ -12,8 +12,8 @@
     },
     {
       "name": "authjs.callback-url",
-      "value": "http%3A%2F%2Flocalhost%3A3000%2Flogin",
-      "domain": "localhost",
+      "value": "http%3A%2F%2Flocalhost%3A3000",
+      "domain": "127.0.0.1",
       "path": "/",
       "expires": -1,
       "httpOnly": true,
@@ -22,14 +22,24 @@
     },
     {
       "name": "authjs.session-token",
-      "value": "eyJhbGciOiJkaXIiLCJlbmMiOiJBMjU2Q0JDLUhTNTEyIiwia2lkIjoicm9WX3hDaFUyS3ZSSFdCN2ZJa1k1eUpGWTlKWHdCSE1VZTBvbU1MdWk1X0F5TVdXOTJveUtDLWFzdml4cTRHWndjOGlHdmtUaGdoQ2lJN3d6X1hvdWcifQ..fkNyGMM8JZbh1JLx5tnk9A.gjxlus2aBQ_vO4pUDgsr6-37iVNuF0iNpEQooi7oYLtuHlOWkeybApqfeLIpWSTnE65tlnSOTxveJkO0YEhsM18uysr475TkoNBp4O3VBWoUF5umGS8VrQYHSkV1vmo0OX9AO_AQLD5ei4qVr7QdT3oYsG0YogqIt-9kKXknj07mDcv6WgWw95Izp4h8r33jxAtwLi8hM8V-3E9EOnXjYOGTIsFcOQ03mRLrGyTmv32kZE5ZdYnDvLPKEkocnKVC201QsRkGyiaBiDaDLhJ_kw.r22gnGgzB0E3AvBAiy4aFoxpIK7IM2OsTb7fEaHdm7k",
-      "domain": "localhost",
+      "value": "eyJhbGciOiJkaXIiLCJlbmMiOiJBMjU2Q0JDLUhTNTEyIiwia2lkIjoicm9WX3hDaFUyS3ZSSFdCN2ZJa1k1eUpGWTlKWHdCSE1VZTBvbU1MdWk1X0F5TVdXOTJveUtDLWFzdml4cTRHWndjOGlHdmtUaGdoQ2lJN3d6X1hvdWcifQ..9w-fr_MsfebWA1GEoqZyWw.ZfgeYI1N82u2lyuDsch5twV-I92NwD18pAsE-duMrYLfs-xKISeJNjd7O5NjStdU_mp-p1XfxMlo9_X6TldXItCQUsNp7mVrzvCviD_s0LnszCB8Wdt3_dXqQan2vPYZ_da3L0WxZsJuO8qq-0PEzg2_2dff4rKSVgl4PHtsO9SdFSicvJye-Bg4BndDA5OXe73-Bmqs0mZyefoQWBtyoiGu2TqzdIiIz2fwFUwV07cXWiBJ7QOpWt5puc9i4UYmUzGKW-Ve3A9fdYrGLef6CQ.yRL7sAH8ZKMn-ZdwQGSmXLtr33gmgGDKGejD1NdKrM4",
+      "domain": "127.0.0.1",
       "path": "/",
-      "expires": 1782133022.888394,
+      "expires": 1782209267.487941,
       "httpOnly": true,
       "secure": false,
       "sameSite": "Lax"
     }
   ],
-  "origins": []
+  "origins": [
+    {
+      "origin": "http://127.0.0.1:3005",
+      "localStorage": [
+        {
+          "name": "SS_SELECTED_PLATFORMS",
+          "value": "[\"cmpi0k2ns00009b2bjqwtjwd9\",\"cmpi0k2ns00019b2bshc6mt07\",\"local-dev-1\",\"local-dev-2\",\"local-dev-3\"]"
+        }
+      ]
+    }
+  ]
 }
