+++ .gemini/agents/dev-agent.md
@@ -3,8 +3,8 @@ name: dev-agent
 description: High-seniority autonomous developer agent. Implements features and fixes bugs.
 kind: local
 tools: ["*"]
-model: gemini-3.1-pro-preview
-# model: gemini-3-flash-preview
+# model: gemini-3.1-pro-preview
+model: gemini-3-flash-preview
 # model: gemini-3.1-flash-lite-preview
 ---
