+++ .gemini/agents/discovery-agent.md
@@ -3,8 +3,8 @@ name: discovery-agent
 description: Senior Solution Architect. Analyzes repo context to turn vague tickets into actionable technical specs.
 kind: local
 tools: ["*"]
-model: gemini-3.1-pro-preview
-# model: gemini-3-flash-preview
+# model: gemini-3.1-pro-preview
+model: gemini-3-flash-preview
 # model: gemini-3.1-flash-lite-preview
 ---
