+++ .gemini/agents/qa-agent.md
@@ -3,8 +3,8 @@ name: qa-agent
 description: Expert Lead QA Automation Writer and Execution Engineer. Designs exhaustive scenarios and writes detailed step-by-step tests.
 kind: local
 tools: ["*"]
-model: gemini-3.1-pro-preview
-# model: gemini-3-flash-preview
+# model: gemini-3.1-pro-preview
+model: gemini-3-flash-preview
 # model: gemini-3.1-flash-lite-preview
 ---
