+++ .gemini/agents/review-agent.md
@@ -3,8 +3,8 @@ name: review-agent
 description: Senior QA and Security Reviewer. Performs deep diff analysis and security audits.
 kind: local
 tools: ["*"]
-model: gemini-3.1-pro-preview
-# model: gemini-3-flash-preview
+# model: gemini-3.1-pro-preview
+model: gemini-3-flash-preview
 # model: gemini-3.1-flash-lite-preview
 ---
