+++ .gemini/state/ticket-565.json
@@ -1,0 +1,15 @@
+{
+  "ticket_id": 565,
+  "ticket_goal": "Refactor: Modularize History Domain (Page & Actions)",
+  "ticket_description": "### Context\nThe History domain has accumulated significant technical debt, with two major files exceeding the project's modularity standards:\n- `src/app/history/page.tsx` (940 lines)\n- `src/app/actions/history.ts` (449 lines)\n\n### Problem/Goal\nThese files are \"God Files\" that violate the 50-line modularity rule. They combine data orchestration, complex UI filtering, and backend logic, making them difficult to maintain, test, and extend.\n\n### Suggested Approach\nPerform a domain-wide refactor to decompose these files into smaller, specialized modules:\n- **Hook Extraction:** Create a `useHistory` hook in `src/hooks/` to manage filtering, pagination, and data synchronization logic.\n- **Component Decomposition:** Break down `history/page.tsx` into smaller MUI-based components (e.g., `HistoryFilters`, `HistoryCard`, `HistoryPagination`).\n- **Action Refactoring:** Modularize `actions/history.ts` by grouping related logic (e.g., `fetchHistory`, `deleteHistoryItem`, `exportHistory`) into separate files or a more granular structure.\n\n### Impact\n- Improves maintainability and developer velocity.\n- Aligns with the project's '50-Line Rule'.\n- Increases testability of individual UI and logic components.",
+  "branch_name": "565-refactor-modularize-history-domain-page-actions",
+  "last_agent": "discovery-agent",
+  "next_agent": "dev-agent",
+  "last_updated_at": "2024-10-24T12:00:00Z",
+  "round-1": {
+    "discovery-agent": {
+      "TECHNICAL_SPECS": "1. **Socratic Audit:**\n   - **Advocate:** Breaking down the 940-line `history/page.tsx` and 449-line `actions/history.ts` into specialized modules radically improves developer experience, aligns with the 50-line modularity rule, and makes individual UI components (e.g., `<HistoryCard>`) reusable.\n   - **Skeptic:** Heavy refactoring risks breaking complex background states, specifically the optimistic UI updates (`cancelledIds`, `processingIds`) and the `usePolling` hook synchronization. Also, Next.js server actions can have bundler issues if dynamic imports (`await import(...)`) are not updated correctly.\n   - **Synthesis:** The refactor is absolutely required to manage technical debt. We will mitigate risks by extracting all state management into a robust `useHistory.ts` hook and carefully updating all dynamic imports across the application.\n\n2. **Action Refactoring (`src/app/actions/history/`)**:\n   Delete `src/app/actions/history.ts` and create a `history` directory with:\n   - `core.ts`: `savePostHistory`, `upsertPlatformResult`, `upsertPlatformResultInternal`.\n   - `retry.ts`: `retryUploadAction`.\n   - `cancel.ts`: `cancelPlatformUploadAction`, `cancelPlatformByPostAction`, `cancelAllUploadsAction`.\n   - `schedule.ts`: `getUpcomingPosts`, `updateScheduledPost`, `publishNowAction`, `deleteScheduledPost`.\n   - `metadata.ts`: `saveStagedMetadata`, `updatePlatformResultsAction`.\n   Update consumers (`src/app/schedule/page.tsx`, `src/lib/actions/ai-chat.ts`, `src/components/dashboard/SidebarInfo.tsx`, etc.) to import from these specific files.\n\n3. **Hook Extraction (`src/hooks/useHistory.ts`)**:\n   Extract state and business logic from `history/page.tsx`.\n   - Manage `posts`, `pendingPost`, `isLoading`, `searchQuery`, `cancelledIds`, `processingIds`.\n   - Encapsulate `fetchHistory`, `handleLoadMore`, `handleRetry`, `handleCancelPlatform`, `handleCancelAll`.\n   - Expose the Cockpit orchestration functions (`executeCockpitDistribution`, `handleCockpitStart`).\n\n4. **Component Decomposition (`src/components/history/`)**:\n   - `HistoryCard.tsx`: Displays an individual post history entry, its metadata, and maps over platforms.\n   - `PlatformResultItem.tsx`: Renders individual platform status indicators and actions (retry/cancel).\n   - `HistoryList.tsx`: Handles mapping and empty states.\n   - `HistoryPage` (`src/app/history/page.tsx`): Becomes a lean orchestrator (< 50 lines) that calls `useHistory()` and renders the components.\n\n5. **Production Readiness:**\n   - Maintain Zod validation schemas within the new server action files.\n   - Ensure Sentry logging (`logger.error`) remains intact during extraction.\n   - Preserve `data-testid` props on UI components for Playwright E2E tests.",
+      "expected_output": "1. `src/app/actions/history.ts` is deleted and replaced by smaller files in `src/app/actions/history/`.\n2. `src/app/history/page.tsx` is significantly reduced in size.\n3. New hooks (`useHistory.ts`) and components (`HistoryCard.tsx`, etc.) are created.\n4. `npx tsc --noEmit` and `npm run build` succeed with no errors, confirming all dynamic and static imports are correct."
+    }
+  }
+}
