+++ src/components/history/HistoryHeader.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import styles from '@/app/history/history.module.css';
+
+interface HistoryHeaderProps {
+  fileInputRef: React.RefObject<HTMLInputElement | null>;
+}
+
+export function HistoryHeader({ fileInputRef }: HistoryHeaderProps) {
+  return (
+    <div className={styles.header}>
+      <div className={styles.headerTop}>
+        <h1 className={styles.title}>Activity Hub</h1>
+      </div>
+      <p className={styles.subtitle}>
+        Track and manage your video distribution in real-time.
+      </p>
+      <input type="file" ref={fileInputRef} style={{ display: 'none' }} accept="video/*" />
+    </div>
+  );
+}
