+++ src/hooks/useHistory.ts
@@ -1,0 +1,1 @@
+export * from './useHistory/index';
